Extract mass property logic into `MassPropertyPlugin`
